tests/wpt/tests/html/canvas/element/the-canvas-state/2d.state.saverestore.globalCompositeOperation.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.state.saverestore.globalCompositeOperation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">

<h1>2d.state.saverestore.globalCompositeOperation</h1>

<script>

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 300;
  canvas.height = 150;
  const ctx = canvas.getContext('2d');

  const old = ctx.globalCompositeOperation;
  ctx.save();
  ctx.globalCompositeOperation = "copy";
  ctx.restore();
  _assertSame(ctx.globalCompositeOperation, old, "ctx.globalCompositeOperation", "old");
}, "save()/restore() restores globalCompositeOperation, for a canvas of size (300, 150).");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 300;
  canvas.height = 150;
  const ctx = canvas.getContext('2d');

  ctx.globalCompositeOperation = "copy";
  const old = ctx.globalCompositeOperation;
  // We're not interested in failures caused by get(set(x)) != x (e.g.
  // from rounding), so compare against `old` instead of "copy".
  ctx.save();
  _assertSame(ctx.globalCompositeOperation, old, "ctx.globalCompositeOperation", "old");
  ctx.restore();
}, "save() does not modify globalCompositeOperation, for a canvas of size (300, 150).");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 0;
  canvas.height = 0;
  const ctx = canvas.getContext('2d');

  const old = ctx.globalCompositeOperation;
  ctx.save();
  ctx.globalCompositeOperation = "copy";
  ctx.restore();
  _assertSame(ctx.globalCompositeOperation, old, "ctx.globalCompositeOperation", "old");
}, "save()/restore() restores globalCompositeOperation, for a canvas of size (0, 0).");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 0;
  canvas.height = 0;
  const ctx = canvas.getContext('2d');

  ctx.globalCompositeOperation = "copy";
  const old = ctx.globalCompositeOperation;
  // We're not interested in failures caused by get(set(x)) != x (e.g.
  // from rounding), so compare against `old` instead of "copy".
  ctx.save();
  _assertSame(ctx.globalCompositeOperation, old, "ctx.globalCompositeOperation", "old");
  ctx.restore();
}, "save() does not modify globalCompositeOperation, for a canvas of size (0, 0).");

</script>
